requirements.txt
dj-database-url==0.4.2
Django==1.8.18
django-crispy-forms==1.4.0
django-durationfield==0.5.1
djangorestframework==3.6.3
gunicorn==19.7.1
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
python-dateutil==2.4.2
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
wrapt==1.10.10
